refactor: extract inspection and observation runtime
