+++ scripts/build_antigravity_distribution.py
@@ -15,13 +15,16 @@ HOOK_FILES = (
     "__init__.py",
     "click_browser.py",
     "click_browser_advisory.py",
+    "click_capability.py",
     "click_verification_policy.py",
     "click_verification_meter.py",
     "click_contract.py",
     "click_dependency_cache.py",
     "click_evidence.py",
     "click_host_coverage.py",
+    "click_inspection.py",
     "click_mutation.py",
+    "click_observation.py",
     "click_process.py",
     "click_service.py",
     "click_state.py",
