+++ tests/repository_policy_core.py
@@ -506,12 +506,15 @@ class RepositoryPolicyTests(unittest.TestCase):
             for name in (
                 "click_browser.py",
                 "click_browser_advisory.py",
+                "click_capability.py",
                 "click_contract.py",
                 "click_dependency_cache.py",
                 "click_evidence.py",
                 "click_host_coverage.py",
+                "click_inspection.py",
                 "click_gate.py",
                 "click_mutation.py",
+                "click_observation.py",
                 "click_process.py",
                 "click_service.py",
                 "click_state.py",
